refactor(cli): 💥 ♻️ rename to github.com/jeffmay/starq to publish properly
🚚 move starq to github.com/jeffmay/starq
🚚 move starq/internal/starq to
   github.com/jeffmay/starq/internal/pkg/starq
🚚 move starq/internal/iox to github.com/jeffmay/starq/internal/pkg/iox

BREAKING CHANGE: CLI package renamed to github.com/jeffmay/starq
