+++ internal/pkg/iox/writers_test.go
@@ -3,9 +3,10 @@ package iox_test
 import (
 	"errors"
 	"io"
-	"starq/internal/iox"
 	"strings"
 	"testing"
+
+	"github.com/jeffmay/starq/internal/pkg/iox"
 
 	"github.com/stretchr/testify/require"
 )
